improve seperation in schema folder
